backend/requirements.txt
# --- Web API ---
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# --- DB / Auth ---
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# --- LangChain stack (SAME LINE: 0.3.x core) ---
langchain==0.3.27
langchain-core==0.3.80
langchain-community==0.3.29
langchain-openai==0.3.11
langchain-chroma==0.2.6
chromadb==1.3.7
langchain-anthropic==0.3.22
langchain-cohere==0.3.4

# --- SDKs ---
openai>=1.50.0,<2
cohere>=5.18.0,<6

# --- Settings / Utils ---
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.0.0
PyYAML==6.0.1
watchdog==3.0.0

# --- Document ingestion ---
PyMuPDF==1.23.8
python-docx==1.1.0
docx2txt==0.8
unstructured==0.11.8
pdfplumber==0.11.8
pytesseract==0.3.13
python-magic==0.4.27
chardet==5.2.0
markdown==3.5.1
Pillow==12.0.0

# --- Data / retrieval ---
openpyxl==3.1.2
pandas==2.1.4
rank-bm25==0.2.2
tiktoken==0.12.0

# --- Guardrail: avoid numpy 2.x surprises ---
numpy<2
